% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_potential.R
\name{date_to_day_numbers}
\alias{date_to_day_numbers}
\title{Convert Date Range to Day Numbers}
\usage{
date_to_day_numbers(start_date, end_date, year)
}
\arguments{
\item{start_date}{Date or character. Start date (e.g., "2025-06-01" or as.Date("2025-06-01"))}

\item{end_date}{Date or character. End date (e.g., "2025-08-31" or as.Date("2025-08-31"))}

\item{year}{Numeric. Year for the calculation}
}
\value{
List with day_start and day_end
}
\description{
Converts start and end dates to day of year numbers
}
\keyword{internal}
