% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_potential.R
\name{set_vostokr_threads}
\alias{set_vostokr_threads}
\title{Set VostokR OpenMP Thread Count}
\usage{
set_vostokr_threads(n_threads = NULL, verbose = FALSE)
}
\arguments{
\item{n_threads}{Integer. Number of threads to use. If NULL, auto-detect based on available cores and lidR settings.}

\item{verbose}{Logical. Print informational messages (default: FALSE)}
}
\value{
No return value, called for side effects (sets thread count).
}
\description{
Control the number of OpenMP threads used by VostokR calculations
}
