% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_normals.R
\name{add_normals}
\alias{add_normals}
\title{Add Normal Vectors to Point Cloud}
\usage{
add_normals(las, k = 10)
}
\arguments{
\item{las}{LAS object from lidR package}

\item{k}{Number of neighbors to use for normal estimation (default: 10)}
}
\value{
LAS object with added normal vectors (nx, ny, nz)
}
\description{
This function adds normal vectors to a LiDAR point cloud using k-nearest neighbors.
The normals are computed using the covariance matrix of neighboring points.
}
\examples{
\dontrun{
library(lidR)
library(vostokR)

# Load test data
LASfile <- system.file("extdata", "test.laz", package="vostokR")
las <- readLAS(LASfile)

# Add normals using 10 nearest neighbors
las <- add_normals(las, k = 10)

# Check that normals were added
names(las@data)
}
}
\seealso{
\code{\link{calculate_solar_potential}} for solar potential calculation
}
