% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_potential.R
\name{get_vostokr_performance_info}
\alias{get_vostokr_performance_info}
\title{Get VostokR Performance Information}
\usage{
get_vostokr_performance_info()
}
\value{
List with OpenMP status and thread information
}
\description{
Get VostokR Performance Information
}
