% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_potential.R
\name{get_vostokr_threads}
\alias{get_vostokr_threads}
\title{Get Current VostokR Thread Count}
\usage{
get_vostokr_threads()
}
\value{
Integer. Current number of OpenMP threads
}
\description{
Get Current VostokR Thread Count
}
