% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_potential.R
\name{plot_solar_potential}
\alias{plot_solar_potential}
\title{Plot Solar Potential Point Cloud}
\usage{
plot_solar_potential(las, ...)
}
\arguments{
\item{las}{LAS object with solar potential values}

\item{...}{Additional arguments passed to lidR::plot()}
}
\description{
Plots solar potential values directly on the point cloud using lidR's native plotting
}
\examples{
\dontrun{
# Plot solar potential on point cloud
plot(las_solar, color = "solar_potential", pal = heat.colors(100))
}
}
